python: moved ecl_grav, ecl_gracl_calc and ecl_subsidence to ecl.gravimetry.
gravimetry tied to ecl.gravimetry.

...
