+++ python/ecl/CMakeLists.txt
@@ -9,6 +9,7 @@ add_subdirectory(eclfile)
 add_subdirectory(grid)
 add_subdirectory(faults)
 add_subdirectory(rft)
+add_subdirectory(gravimetry)
 #add_subdirectory(ecl)
 add_subdirectory(geo)
 add_subdirectory(test)
