+++ python/ecl/gravimetry/ecl_grav_calc.py
